Fix the chooser and behaviour for networks
