+++ src/networks/wagtail_hooks.py
@@ -3,6 +3,7 @@ from wagtail import hooks
 from wagtail.admin import widgets as wagtailadmin_widgets
 
 from networks.models import Network, NetworkContentPage
+from networks.views import NetworkChooserViewSet
 
 
 @hooks.register("register_page_listing_more_buttons")
@@ -29,3 +30,8 @@ def page_listing_more_buttons(page, page_perms, is_parent=False):
             ),
             priority=10,
         )
+
+
+@hooks.register("register_admin_viewset")
+def register_network_chooser_viewset():
+    return NetworkChooserViewSet("network_chooser", url_prefix="network-chooser")
