+++ src/peoplefinder/wagtail_hooks.py
@@ -1,7 +1,6 @@
 from wagtail import hooks
 
 from peoplefinder.views.chooser import (
-    NetworkChooserViewSet,
     PersonChooserViewSet,
     TeamChooserViewSet,
 )
@@ -15,8 +14,3 @@ def register_person_chooser_viewset():
 @hooks.register("register_admin_viewset")
 def register_team_chooser_viewset():
     return TeamChooserViewSet("team_chooser", url_prefix="team-chooser")
-
-
-@hooks.register("register_admin_viewset")
-def register_network_chooser_viewset():
-    return NetworkChooserViewSet("network_chooser", url_prefix="network-chooser")
