+++ src/peoplefinder/widgets.py
@@ -1,7 +1,7 @@
 from django.utils.translation import gettext_lazy as _
 from generic_chooser.widgets import AdminChooser
 
-from peoplefinder.models import Network, Person, Team
+from peoplefinder.models import Person, Team
 
 
 class PersonChooser(AdminChooser):
@@ -20,12 +20,3 @@ class TeamChooser(AdminChooser):
     model = Team
     choose_modal_url_name = "team_chooser:choose"
     icon = "group"
-
-
-class NetworkChooser(AdminChooser):
-    choose_one_text = _("Choose a network")
-    choose_another_text = _("Choose another network")
-    link_to_chosen_text = _("Edit this network")
-    model = Network
-    choose_modal_url_name = "network_chooser:choose"
-    icon = "globe"
